refactoring: user & item
Code Refactor of User and Item Models
