+++ backend/app/alembic/env.py
@@ -18,8 +18,9 @@ fileConfig(config.config_file_name)
 # target_metadata = mymodel.Base.metadata
 # target_metadata = None
 
-from app.models import SQLModel  # noqa
+from sqlmodel import SQLModel  # Direkter Import aus dem sqlmodel Paket
 from app.core.config import settings # noqa
+
 
 target_metadata = SQLModel.metadata
