+++ backend/app/api/routes/items.py
@@ -5,7 +5,8 @@ from fastapi import APIRouter, HTTPException
 from sqlmodel import func, select
 
 from app.api.deps import CurrentUser, SessionDep
-from app.models import Item, ItemCreate, ItemPublic, ItemsPublic, ItemUpdate, Message
+from app.models.user import Message
+from app.models.item import Item, ItemCreate, ItemPublic, ItemsPublic, ItemUpdate
 
 router = APIRouter()
