+++ backend/app/crud.py
@@ -4,7 +4,11 @@ from typing import Any
 from sqlmodel import Session, select
 
 from app.core.security import get_password_hash, verify_password
-from app.models import Item, ItemCreate, User, UserCreate, UserUpdate
+
+# Importiere die Modelle aus den entsprechenden Dateien
+from app.models.user import User, UserCreate, UserUpdate
+from app.models.item import Item, ItemCreate
+
 
 
 def create_user(*, session: Session, user_create: UserCreate) -> User:
