+++ backend/app/models/__init__.py
@@ -1,0 +1,13 @@
+from .user import (
+    User,
+    UserCreate,
+    UserUpdate,
+    Message,
+    UpdatePassword,
+    UserPublic,
+    UserRegister,
+    UsersPublic,
+    UserUpdateMe
+)
+
+from .item import Item
